requirements.txt
docx==0.2.4
elasticsearch==8.10.0
fitz==0.0.1.dev2
Flask==2.2.5
Flask_Cors==4.0.0
openpyxl==3.1.2
pandas==2.0.2
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.0
Pillow==10.4.0
psycopg2_binary==2.9.9
pytest==8.2.0
python_docx==0.8.11
python_pptx==0.6.22
Requests==2.32.3
scikit_learn==1.3.0
scikit_learn==1.2.2
spacy==3.5.3
